% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dheatmap.R
\name{Dheatmap}
\alias{Dheatmap}
\title{Plot Hoeffding's D statistic against every other sample}
\usage{
Dheatmap(...)
}
\arguments{
\item{...}{Output from Dall() performed on one or more log transformed matrices, with columns corresponding to samples and rows to genes or other features.}
}
\description{
This function takes the output of Dall, with the values off Hoeffding's D for each sample compared to every other sample in each expression matrix, and gives a heatmap of D statistics.
}
\examples{
Dstat_results_all = Dall(expr, expr.batchcorrected, expr.normalised)
Dheatmap(Dstat_results_all)
}
