% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPlot.R
\name{Dmedian}
\alias{Dmedian}
\title{Hoeffding's D statistic against the median for an expression matrix}
\usage{
Dmedian(...)
}
\arguments{
\item{...}{One or more log transformed matrices, with columns corresponding to samples and rows to genes or other features.}
}
\description{
This function allows you to take one or more log-transformed expression matrices, and calculates Hoeffding's D-statistic for each sample against the median for that expression matrix. This outputs a dataframe with the values off Hoeffding's D for each sample in each expression matrix.
}
\examples{
Dmedian(expr)
DstatResults = Dmedian(expr, expr.batchcorrected, expr.normalised)
}
