% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPlot.R
\name{Dplot}
\alias{Dplot}
\title{Plot Hoeffding's D statistic against the median for an expression matrix}
\usage{
Dplot(Dmedian_output)
}
\arguments{
\item{Dmedian_output}{The output after having performed Dmedian() on one or more log transformed matrices, with columns corresponding to samples and rows to genes or other features.}
}
\description{
Generates a violin plot from the output of Dmedian to quickly visualise the distribution of D statistics.
}
\examples{
DstatResults = Dmedian(expr, expr.batchcorrected, expr.normalised)
Dplot(DstatResults)
}
