% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAplot.R
\name{MAplot}
\alias{MAplot}
\title{MA plot}
\usage{
MAplot(expression, sample1, sample2 = NULL)
}
\arguments{
\item{expression}{A log-transformed expression matrix, with columns corresponding to samples and rows to genes or other features.}

\item{sample1}{Character string specifiying the name of a column in your expression matrix, corresponding to a sample}

\item{sample2}{Character string specifiying the name of a column in your expression matrix, corresponding to a sample. Defaults to the median if not specified.}
}
\description{
This function allows you to take a log-transformed expression matrix, and specify one (in which case it will be compared to the median) or two samples, outputing an MA plot, where A = mean expression count or intensity and M = log fold difference
}
\examples{
data(expr.raw, expr.batchcorrected, expr.normalised)
MAplot(expr.raw, "S8")
MAplot(expr.raw, "S8", "S15")
MAplot(expr.normalised, "S8", "S15")
}
