% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dheatmap.R
\name{sweetDall}
\alias{sweetDall}
\title{Hoeffding's D statistic against every other sample}
\usage{
sweetDall(...)
}
\arguments{
\item{...}{One or more log transformed matrices, with columns corresponding to samples and rows to genes or other features.}
}
\description{
This function allows you to take one or more log-transformed expression matrices, and calculates Hoeffding's D-statistic for each sample in the matrix against every other sample. This outputs a dataframe with the values off Hoeffding's D for each sample compared to every other sample in each expression matrix.
}
\examples{
data(expr.raw, expr.batchcorrected, expr.normalised)
sweetDall(expr.raw)
Dstat_results_all = sweetDall(expr.raw, expr.batchcorrected, expr.normalised)
}
