% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dheatmap.R
\name{sweetDheatmap}
\alias{sweetDheatmap}
\title{Plot Hoeffding's D statistic against every other sample}
\usage{
sweetDheatmap(Dall_output)
}
\arguments{
\item{Dall_output}{Output from sweetDall() performed on one or more log transformed matrices, with columns corresponding to samples and rows to genes or other features.}
}
\description{
This function takes the output of sweetDall, with the values off Hoeffding's D for each sample compared to every other sample in each expression matrix, and gives a heatmap of D statistics.
}
\examples{
data(expr.raw, expr.batchcorrected, expr.normalised)
Dstat_results_all = sweetDall(expr.raw, expr.batchcorrected, expr.normalised)
sweetDheatmap(Dstat_results_all)
}
