% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPlot.R
\name{sweetDmedian}
\alias{sweetDmedian}
\title{Hoeffding's D statistic against the median for an expression matrix}
\usage{
sweetDmedian(...)
}
\arguments{
\item{...}{One or more log transformed matrices, with columns corresponding to samples and rows to genes or other features.}
}
\description{
This function allows you to take one or more log-transformed expression matrices, and calculates Hoeffding's D-statistic for each sample against the median for that expression matrix. This outputs a dataframe with the values of Hoeffding's D for each sample in each expression matrix.
}
\examples{
data(expr.raw, expr.batchcorrected, expr.normalised)
sweetDmedian(expr.raw)
DstatResults = sweetDmedian(expr.raw, expr.batchcorrected, expr.normalised)
}
