% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPlot.R
\name{sweetDplot}
\alias{sweetDplot}
\title{Plot Hoeffding's D statistic against the median for an expression matrix}
\usage{
sweetDplot(Dmedian_output)
}
\arguments{
\item{Dmedian_output}{The output after having performed Dmedian() on one or more log transformed matrices, with columns corresponding to samples and rows to genes or other features.}
}
\description{
Generates a violin plot from the output of sweetDmedian to quickly visualise the distribution of D statistics.
}
\examples{
data(expr.raw, expr.batchcorrected, expr.normalised)
DstatResults = sweetDmedian(expr.raw, expr.batchcorrected, expr.normalised)
sweetDplot(DstatResults)
}
